Move logo into own folder. Saves getting too messy
